requirements.txt
azure-core==1.39.0
azure-eventhub==5.11.7
azure-iot-device==2.14.0
azure-storage-blob==12.28.0
blinker==1.8.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.1.7
colorama==0.4.6
cryptography==46.0.6
deprecation==2.1.0
Flask==3.0.3
Flask-Cors==4.0.0
gunicorn==23.0.0
idna==3.11
isodate==0.7.2
itsdangerous==2.2.0
janus==2.0.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
paho-mqtt==1.6.1
pycparser==3.0
PySocks==1.7.1
python-dotenv==1.2.2
requests==2.33.0
requests-unixsocket2==1.0.1
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.0.4
